// components/Preview/types.ts
export enum MultilineBlockType {
  CODE = 'mbt_code',
}

export enum ListBlockType {
  ORDERED_LIST = 'lbt_orderedList',
  UNORDERED_LIST = 'lbt_unorderedList',
  QUOTE = 'lbt_quote',
}

export enum InlineMarkdownType {
  CODE = 'imt_code',
  LINK = 'imt_link',
}

export enum LineType {
  DEFAULT = 'lt_default',
  H1 = 'lt_h1',
  H2 = 'lt_h2',
  H3 = 'lt_h3',
  H4 = 'lt_h4',
  H5 = 'lt_h5',
  H6 = 'lt_h6',
}